Renamed Casting_2 => Set_movable_separability_2
