+++ Documentation/doc/Documentation/packages.txt
@@ -23,7 +23,7 @@ h1 {
 
 \package_listing{Matrix_search}
 \package_listing{QP_solver}
-\package_listing{Casting_2}
+\package_listing{Set_movable_separability_2}
 
 \section PartKernels Geometry Kernels
