+++ Set_movable_separability_2/doc/Set_movable_separability_2/Doxyfile.in
@@ -1,3 +1,3 @@
 @INCLUDE = ${CGAL_DOC_PACKAGE_DEFAULTS}
 
-PROJECT_NAME = "CGAL ${CGAL_DOC_VERSION} - 2D Castings"
+PROJECT_NAME = "CGAL ${CGAL_DOC_VERSION} - 2D Movable Separability of Sets"
