+++ Set_movable_separability_2/doc/Set_movable_separability_2/PackageDescription.txt
@@ -1,17 +1,17 @@
-/// \defgroup PkgCasting2 2D Castings Reference
+/// \defgroup PkgSetMovableSeparability2 2D Castings Reference
 
-/// \defgroup PkgCasting2Funcs Free Functions
-/// \ingroup PkgCasting2
+/// \defgroup PkgSetMovableSeparability2Funcs Free Functions
+/// \ingroup PkgSetMovableSeparability2
 
 /*!
-\addtogroup PkgCasting2
+\addtogroup PkgSetMovableSeparability2
 \todo check generated documentation
-\cgalPkgDescriptionBegin{2D Castings,PkgCasting2Summary}
+\cgalPkgDescriptionBegin{2D Castings,PkgSetMovableSeparability2Summary}
 \cgalPkgPicture{Casting_2.png}
 \cgalPkgSummaryBegin
 \cgalPkgAuthors{Shahar Shamai, Efi Fogel}
 \cgalPkgDesc{This package consists of the implementations of various predicates and operations related to castings of polygonal objects. In particular, it can be used to determine whether a mold for a polygonal object does exist. If a mold exists, the package can also be used to compute all possible orientations of the feasible molds and the corresponding motions needed to remove the casted object from the mold.}
-\cgalPkgManuals{Chapter_Castings_of_polygons,PkgCasting2}
+\cgalPkgManuals{Chapter_Castings_of_polygons,PkgSetMovableSeparability2}
 \cgalPkgSummaryEnd
 \cgalPkgShortInfoBegin
 \cgalPkgSince{4.10}
