+++ Set_movable_separability_2/doc/Set_movable_separability_2/examples.txt
@@ -1,3 +1,3 @@
 /*!
-\example Casting_2/single_mold_translational_casting.cpp
+\example Set_movable_separability_2/single_mold_translational_casting.cpp
 */
